Use Java standard SHA3 implementation (#4459)
We get a lot of noisy security alerts about a dependency on
Bouncycastle, which was introduced historically to work around the lack
of a SHA3-256 implementation in standard Java. This PR removes the
dependency by porting its single usage (rule digests / `UNIQUE_ID`s)
over to use new features in the standard library.

The IDs produced after this change are identical to their previous form.