add two important missing flags to the llvm backend compiler (#4472)
This PR fixes a small fraction of a significant piece of technical debt
in pyk's implementation of `pyk.ktool.kompile`, adding support for two
command line flags that are important that were currently unable to be
passed to the compiler due to its design. This is by no means an
exhaustive list. There are still other important flags that are
unsupported. This is simply what is likely to block me in the immediate
future.

Co-authored-by: Bruce Collie <brucecollie82@gmail.com>
Co-authored-by: rv-jenkins <admin@runtimeverification.com>