Fix bug report request IDs being reset when using multiple `KoreClient`s (#4480)
Fixes #4478.

We were generating response json files for bug reports in a non-unique
way so that when we have multiple `JsonRpcClient`s they will overwrite
each other's response files. This change ensures that each client will
have a unique ID associated which will be included in the path of the
response files generated by that client so there are no more name
conflicts.

Tested generating bug reports for `kontrol prove` and `kevm prove`.

---------

Co-authored-by: rv-jenkins <admin@runtimeverification.com>
Co-authored-by: Tamás Tóth <tothtamas28@users.noreply.github.com>