Deprecate `symbol` and `klabel` (#4045)
This PR implements @tothtamas28's warning messages for usages of the
deprecated `klabel(_)` and `symbol` attributes (which can now be
unambiguously replaced by `symbol(_)` and `overload(_)`).

The actual code changes are pretty minimal, but we need to do a bunch of
updates on test cases and related output to keep tests passing.
Additionally, we have pre-merged downstream PRs to make sure other
projects are not subjected to noisy compiler output as a result of
making this change:

Downstream updates:
- [x] C https://github.com/runtimeverification/c-semantics/pull/1125
- [x] EVM https://github.com/runtimeverification/evm-semantics/pull/2378
- [x] WASM
https://github.com/runtimeverification/wasm-semantics/pull/666
- [x] MX Backend
https://github.com/runtimeverification/mx-semantics/pull/298
- [x] Kasmer
https://github.com/runtimeverification/kasmer-multiversx/pull/168
- [x] MIR (new semantics doesn't use `klabel`)
- [x] Pyk (this PR, now that K and Pyk are merged)
- [x] Blockchain Plugin
https://github.com/runtimeverification/blockchain-k-plugin/pull/184