install-build-deps: Automate some instructions from the README
This removes the `apt-get install` and `brew install` instructions and
lists of packages (and, for Debian, Haskell Stack as well) from the
README.md file and moves them to a script that will determine which set to
install and do the install. (This can easily be expanded to support other
systems if necessary.) The improvement to developer convenience should be
obvious.

This also tweaks the bits of Markdown we touched to move the referenced
URLs to the end of the file, to make it easier to read the plaintext.

The script has been tested on Debian 12, Bash 3.2.48(3) on a Debian system
and on MacOS (using some version of Bash 3).