Nix refactoring (#4500)
This PR is a preparatory of the K Nix build code that will allow us to
further simplify downstream packages. Roughly, the steps taken here are:
- Move some definitions from the set of K flake output packages into the
K overlay; this lets us reference `k` directly from semantics that use
the overlay, rather than `k-framework.${pkgs.system}.packages.k` as we
do now.
- Pull the bulk of the Pyk flake out into a standalone overlay
expression; the actual build code in that overlay is identical.
- Re-export Pyk from the K overlay using the extracted Pyk overlay.
- Test both ways of referencing Pyk (via subdirectory flake, or via K
attribute) in CI.

Once this PR is merged, the next step will be to replace usages of Pyk
as a standalone flake in downstream code to instead point at the
corresponding K attribute. This simplifies version-bump and Nix input
management in those projects.

I have tested this against the WASM semantics downstream, and the
suggested new method for referencing Pyk successfully builds there.

---------

Co-authored-by: rv-jenkins <admin@runtimeverification.com>