Allow log origin "proxy" (#4520)
With https://github.com/runtimeverification/haskell-backend/pull/3974,
`kore-rpc-booster` increments execution depth and emits a rewrite trace
in the scenario of pruning all but one branches in a branching execute
response. The trace has origin `"proxy"`, because that's where the step
is effectively taken. This PR makes sure `pyk` is ready to handle this.

Co-authored-by: rv-jenkins <admin@runtimeverification.com>