Make request IDs unique across clients (#4524)
Changes the ID fields of the requests being sent to the kore RPC server.
Before they were just integers counting up the requests of each client,
but it is useful to be able to see in the logs which requests are coming
from which client, so this changes the IDs to be this same counter, but
prepended with `id()` of the client that is making the request.

---------

Co-authored-by: Tamás Tóth <tothtamas28@users.noreply.github.com>