Fix crash with failure info with custom multiplicity cells (#4531)
This crash was found by @lucasmt. When there is a structural match
failure in failure info computation we were assuming the cells in the
antecedent were always present in the consequent, but this is not
necessarily true in the case where we have optional cells, e.g.
`type="Map"`. This avoids the crash by simply indicating the missing
cell is missing from the consequent. There is still confusing behavior
when, e.g. there are multiple copies of the same cell.