Further fine-tune `Int2Bytes(Int, Endianness, Signedness)` equations for booster (#4533)
The previous change #4522 was insufficient to make the equations usable
in `booster`. Matches with literal `Int` constants will still be
indeterminate, so this PR removes the literals from the rule LHSs.

* Use `0 ==Int I` instead of a syntactic _match_ in the equation for
argument `0`. This enables evaluating terms with `I ==Int 0` in the path
condition.
* Remove equation with literal `-1` and widen requires in equation for
negative arguments (previously `I <Int -1`, now `I <Int 0`). The result
is the same as before with no extra equation required.