install-build-deps script (#4505)
Add an `install-build-deps` script to replace the manual commands in the
README. See commit message(s) for details.

__WARNING:__ This commit is based on the commit in #4054; if you merge
this it will merge that one too. (The script includes the
`libunwind-dev` package added in that commit.)

You'll probably want to do a quick check that this works on MacOS and
update the commit message appropriately.