Add extra check for lldb path in krun --debugger (#4560)
When invoking `krun --debugger` on MacOS, we check a file that's
supposed to specify where the lldb we use is installed on the system.

Sometimes the contents of this file are `LLDB-NOTFOUND` and we're
supposed to fall back to whatever lldb is on the user's PATH. There's a
bug where we don't handle this case properly, and this change fixes
that.