Status bar support in proof loop (#4561)
Adds a couple things needed for supporting a proof status bar for
real-time updates in kevm/kontrol: Adds a callback parameter to the two
proof loops to get information about the proofs during execution of the
proof loop (without having to read from disk). Also adds a field to
`APRProof` which is the one-line summary which will be printed in the
status bar, along with a default implementation in `Proof`.