Redesigning rewrite logging (#4580)
This PR adds the option to turn on/off the logging of both successful
and failed rewrites. While the latter was already present, the former
was observed to introduce an execution speedup in engagement proofs, and
is therefore introduced.

As part of this PR, the `trace_rewrites` is renamed into
`log_fail_rewrites`, which is both aligned with `log_succ_rewrites` and
describes its purpose better.