dedupe closures generated by #let (#4584)
Previously we were generating one LHS term in each `#let` lambda per use
of a particular variable. This introduced complexity for the resulting
rule which needed to check that each capture of the same variable had
the same value, which is unnecessary work. We fix this by simply
deduping the closure so that each variable can be captured at most once
by any given lambda.