Adjust the sort of `mlEquals` (#4633)
Currently `mlEquals` has `GeneratedTopCell` as the default argument
sort, which is pretty specific/limited. Instead, we set `K` as the
default argument sort (supersort of all sorts), and then go to all the
sites that use it and specify the argument sort more carefully.

This doesn't affect any downstream repos, none of them use `mlEquals`
directly.