More readable and informative status bar (#4636)
This PR adjusts the proof info shown in the status bar. In particular,
it:
- removes the `branches` info, since this number refers to all of the
internal branchings of the proof, it differs from the number of leaves,
and it is not clear how it is useful to the end-user;
- removes the `terminal` info, since this can be computed from the other
information; it could also be confusing to end-users as to why there are
always two terminal nodes to begin with (final node of `setUp` and
`target`);
- reduces horizontal space so that there is a greater chance that full
status does not overflow and more info can be seen when test name is
large.