Impure attribute in KORE (#4637)
We add back code that was previously removed that was designed to
compute which functions in K are impure (i.e., have side effects).

We don't do anything special for these functions except emit an
attribute, but the llvm backend needs this information because it needs
to know these functions cannot be optimized out.

---------

Co-authored-by: rv-jenkins <admin@runtimeverification.com>