Handle list random access patterns in frontend (#4634)
This PR adds support for the `L [ I <- X ]` syntax on the LHS of rules,
which allows you to pattern match on an arbitrary element of a list.

The code is not fully polished; it is not integrated with configuration
abstraction or cell collections. It is not likely to be further improved
upon presently because the code it was intended to support ended up not
proving worth merging. However, it is broadly useful functionality
(along with the accompanying LLVM backend PR) and thus ought to be
merged.

---------

Co-authored-by: Dwight Guth <dwight.guth@runtimeverification.com>