Roll back to pre mmap in llvm backend (0.1.103) (#4713)
This rolls back the LLVM backend dependency to 0.1.105, to work around
problems with the new thread-local `mmap` memory management.

It also rolls back a change that introduced the new
`LLVMFunctionExitEvent` in pyk (to cater for its addition in LLVM
backend), which is in the first commit of this branch (tag
`added-llvm-function-exit`, to retain the commit for later re-addition).

---------

Co-authored-by: rv-jenkins <devops@runtimeverification.com>