Process complex LHS patterns in function rules (#4757)
Eliminates patterns that are not supported by Lean from the LHS of
function rules.

For example, since `_List_` is a function, instead of matching the last
item of a list as

```lean
noncomputable def _2ede380 : SortList → Option SortInt
  | _List_ _Gen0 (ListItem (SortKItem.inj_SortInt I)) => ...
  | _ => none
```

the following definition is generated:

```lean
noncomputable def _2ede380 : SortList → Option SortInt
  | _Pat0 => match (ListHook SortKItem).split _Pat0.coll 0 1 with
    | some ([], _Gen0, [SortKItem.inj_SortInt I]) => ...
    | _ => none
```

where `split` is repsonsible for deconstructing the list into prefix,
middle and suffix.