Refactor adding modules via `CTermSymbolic` (#4771)
This PR factors out the ability to add KAST level modules to the RPC
server via `CTermSymbolic`, so that the user doesn't need to manually do
the conversions needed for adding such modules. The `APRProver` is
refactored to use this new `CTermSymbolic.add_module` as well.

This is part of
https://github.com/runtimeverification/kontrol/issues/977, and blocking
https://github.com/runtimeverification/kontrol/pull/979.