`Prelude.lean`: embed carrier implementation in hook signatures (#4770)
This PR embeds the carrier implementation into the hook signatures in
`Prelude.lean`.

Before #4766, we used to have an abstract carrier field for hook
signatures. However, even instantiating that field with a concrete
implementation clashes with Lean's positivity checker. This occurs in
the `mutual` block definition involving `KItem`.
As a workaround, the hooks now only reflect the signature of the
functions that should be implemented in the future. The carrier is now
also instantiated in the hook signatures.