CTerm.anti_unify: Don't call ml_pred_to_bool unless necessary
Since these values are only used if keep_values is set, we do not need to always
compute them. In addition, it avoids exceptions when the predicated cannot be
converted to a bool expression: e.g. when using #Ceil

I'm not sure why the terms need to be converted into Bool expressions (vs ML
predicates) in the first place. Thus, while this is only a work-around (it would
still break if keep_values=True), it lets us move forward without risking
breaking other projects that rely on the original behaviour.