Allow introducing semantic rules in kprove (#4772)
Part of: https://github.com/runtimeverification/kontrol/issues/977

Currently, `kprove` does not allow including new semantic rules in the
definition that is being proven, only claims and `simplification` rules.
This change allows the user to set a flag `--allow-new-rules` which
disables this check. The intention is for it to be used with
`--dry-run`, so that `kprove ...` can be used just to parse some modules
basically.

In pyk, we add `--allow-new-rules` to `KProve.parse_module(...)`
invocation, so that it can parse new modules that introduce semantic
rules. This enables us to use those new rules in the haskell backend
booster when doing proofs with Kontrol, which is key to loading loop
invariants dynamically to support
https://github.com/runtimeverification/kontrol/issues/977.