`Prelude.lean`: fix error on the length of `Array.?pad` (#4823)
This PR addresses a change not up-streamed in #4822, which implements
(among others) the
`«padRightBytes(_,_,_)_BYTES-HOOKED_Bytes_Bytes_Int_Int»` and
`«padLeftBytes(_,_,_)_BYTES-HOOKED_Bytes_Bytes_Int_Int»` as
`Array.rightpad` and `Array.leftpad` respectively.
However, the length parameter passed to this functions was `Δ :=
len.toNat - b.size`, but it ought to be `len.toNat`, since these
functions pad to the maximum of the length provided.

Co-authored-by: automergerpr-permission-manager[bot] <190534181+automergerpr-permission-manager[bot]@users.noreply.github.com>