Update nix flake inputs structure and add `rv-nix-tools` to dependency update workflow (#4824)
This pull request adapts the nix flake inputs structure regarding
`nixpkgs` and `rv-nix-tools`. In particular, `nixpkgs` is now specified
for inputs and transitive inputs by following an explicitly specified
revision of `rv-nix-tools`. More thorough reasoning for this change is
outlined in documentation added by this PR
runtimeverification/rv-nix-tools#14. This change is done to make
`nixpkgs` properly updatable by CI.

In addition, this pull requests adapts the `update-deps.yml` workflow to
also update `rv-nix-tools` that is used to implicitly specify the
revision of `nixpkgs`.

---------

Co-authored-by: Freeman <105403280+F-WRunTime@users.noreply.github.com>