Fix Dockerfiles for release workflows (#4835)
The `poetry` to `uv` migration pull request #4831 causes the Dockerfile
builds in release workflows [on CI to
fail](https://github.com/runtimeverification/k/actions/runs/15757039533/job/44414666887).
This is due to `uv` not being found at runtime as the underlying `bin`
directory is missing in the PATH variable. This pull request fixes this
by introducing a non-root user whose local bin directory
`/home/${USER}/.local/bin` is added to the PATH variable.

I tested that the Dockerfiles successfully build on my local machine.

In addition, this pull request sets the `UV_VERSION` argument for the
Dockerfile in the `with-k-docker` GitHub action that was missing before,
which caused the resulting uv version that was installed to be of the
latest version instead of a fixed specified version.