Update nix lockfile (#4841)
In PR #4838, the nix lock file `flake.lock` was updated to be in sync
again with `flake.nix`, which is a hard-requirement for `kup` and `nix`
to be able to install `k` properly. This PR also added a check in CI
that makes sure that the lock file is actually in sync with the flake
file.

In PR #4834, the develop branch was merged with the PR branch, which
causes the new check to be included in the PR branch. In addition, the
`update deps` workflow of the respective update PR was automatically run
that updated nix lock file. For some reason, the `flake.lock: update`
commit [generated a wrong lock
file](https://github.com/runtimeverification/k/pull/4834/commits/76f0afe0bfe9bf2677daeb5f0b1b33e499d46084#diff-216b2b7bfde9416c79d133bacb031e95702a20bdedb548c0b055c837aa4f6a9cL169)
that is un-sync with the `flake.nix` file in the same commit. Normally,
this should have been detected by the new check in the test PR workflow,
but it was not. The [first CI test run
succeeded](https://github.com/runtimeverification/k/actions/runs/16276235659/attempts/1)
and the PR was merged afterwards. I restarted a CI run on this PR,
[which failed as
expected](https://github.com/runtimeverification/k/actions/runs/16276235659/job/46028193744#step:7:20).

In my opinion, this should not actually happen and could have been
caused by a checkout of the wrong commit in CI? But on a second thought,
I observed that the [test CI workflow succeeded for
MacOS](https://github.com/runtimeverification/k/actions/runs/16276235659/job/46028193732#step:6:16)
due to the store paths being already in the nix store, which means that
the nix flake was actually build already build by CI as is, without
changes.

So in conclusion, this is either sum bug in the checkout of repositories
(unlikely I guess) or some very weird bug by nix (more likely).
Therefore, we should update the versions of nix used in the MacOS CI and
the workflows of this repository, as both use rather old versions of
nix.