Allow publishing releases without macos-tests (#4853)
This PR allows bypassing the `macos-test` CI job to publish new K
releases.

The release workflow started failing in the `macos-build` job as of this
PR:
https://github.com/runtimeverification/k/pull/4850
The failure seems unrelated to the changes introduced by the PR and
looks more like a flakiness issue in the CI due to non-deterministic
dependency resolution.

A follow-up PR fixed the `macos-build` job:
https://github.com/runtimeverification/k/pull/4851

Unfortunately, the workflow is now failing in the `macos-test` job.
https://github.com/runtimeverification/k/pull/4852 is an attempt to
debug and fix the tests.

The inability to publish new K-releases is currently blocking important
updates on Simbolik.